refactor: gộp giao diện chi tiết đơn hàng
